In this step (2), I made changes in 3 different areas
I created a NEW class called EmployeeMain.This class role is to hold the
main.I found
that without an actual Main class,a person has to skim through all the
classes to know where the main class is located but now a person can
just run the application and understand what is the application about
before jump into understanding the code.

Inside the employee class i removed the tostring method since the
program is ran on an interface.Also inside the EmployeeSummerDialog, I
used a for each loop by replacing the for loop from the Container
summerPane constructor.